config.org
I still need to add some things that got lost with the last windows update.
But for now I need to work on the compiler.
I should really consider cleaning the mess that this repository as become. I should
only be uploading the configuration file itself, without the extra packages. But since
Emacs windows development is not so stable, I don't know If I should.